Correct Functionality of service layer

+++ src/main/java/com/chirag/exceltomysql/repository/ExcelRepository.java
@@ -3,5 +3,5 @@ package com.chirag.exceltomysql.repository;
 import com.chirag.exceltomysql.entity.Orders;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExcelRepository extends JpaRepository<Orders, Integer> {
+public interface ExcelRepository extends JpaRepository<Orders, Long> {
 }
